refactor: move some tests from modulesource_test to  modulerefs_test
